[FIO toup] sks ta: fix build warnings on 64 bits
Fix build warnings when building SKS TA for qemu armv8.

Mainly size_t x uint32_t differences.

+++ ta/secure_key_services/src/processing.c
@@ -434,7 +434,7 @@ uint32_t alloc_get_tee_attribute_data(TEE_ObjectHandle tee_obj,
 {
 	TEE_Result res = TEE_ERROR_GENERIC;
 	void *ptr = NULL;
-	size_t sz = 0;
+	uint32_t sz = 0;
 
 	res = TEE_GetObjectBufferAttribute(tee_obj, attribute, NULL, &sz);
 	if (res != TEE_ERROR_SHORT_BUFFER)
